Copy over template used for Tappy Terror index
Brings over the template used for the Tappy Terorr index page from the web.py
version. The new version still needs to be properly ported to a jinja template
in order to actually do everything the original one did, but this will at
least give a scaffolding to start converting the web layer stuff to flask.