Port Location over
Extracts the Location class from the old Tappy Terror and brings it over to
Flask-land. This was straightforward: the class is self contained and just
encapsulates a bit of logic about mob spawning.

The only hard decision was what to do about the things that were set in module
level variables as configuration in the original Tappy Terror. I've decided to
keep them as module level variables for now just for the ease of porting, but
I think they should go into their own config structure if we end up with
a bunch of them. We didn't change these much when we were setting up the
game initially, but I still find it icky that we had to tune them by
changing code, not data.

Adds tests for mob spawning and removal to ensure that gameplay asusmptions
about mob counts aren't violated.