Port Polygon over
Bring Polygon over in to Flask-land. We used this to represent bounding boxes
for rooms in the game because there obviously isn't a good alternative to
rolling your own class when it comes to representing 2D bounding boxes :)
This was another straightforward section of code to port because it doesn't
depend on anything else.

I'm not sure why this class is around and useful, thanks me from many years
ago. It should have some more behaviors if it's actually going to be more
than "hey here's a vertex list".

Also bring over the floor list, which defines the bounds of various rooms
in the game. This is the primary place that Polygons get constructed,
so it's easy to pull over at the same time.

Adds a couple of silly tests that basically just construct Polygons so
that there's at least a little test coverage.