Initial pass at state snapshot schema
First pass at moving the Tappy Terror state snapshots from disk into the db.
The general plan follows the outline previously in schema.sql.

I've split the data for locations and teams into tables for the data that
does not mutate over the course of a game and _ephemera tables for the data
that could potentially change every snapshot.

The _ephemera tables could use the sqlite 3.8 WITHOUT ROWID optimization if
desired, but the default sqlite3 on my dev system is 3.7.13 so I want it to
work without upgrading sqlite.