Order the bounds_vertices lists
Adds list_rank to bounds_vertices so that the collections can be kept in
order. The vertex lists need to be in order for us to draw them correctly.
We can't actually guarantee that the query order from SQLite is the same as
the insert order, so this extra data helps us out.

This isn't actually normalized because I feel like we should not need to
reorder entries in the vertex lists.