Endpoint for data pushed from OpenAMD
Adds an endpoint '/amd/push' which responds to posts of json data from
OpenAMD with new location data for our players. We expect to recieve data from
OpenAMD as a JSON list of user objects.

This does not implement the game logic which should be triggered on a new
batch of data.