Abort with status 422 on valid json in an unexpected format
Previously I returned a json object with status: "failure" on this error, but
that behavior made me incorrectly assume some requests I made with sim.py
were OK when they were in fact malformed. In sim.py I had just checked if
the status was 200 originally and didn't actually check the return status. It
seems quite likely that other people would do the same, so I'm making this
error more RESTful.

I decided to use error code 422 Unprocessable Entity, stealing a page from
WebDav even though this isn't WebDav. Twitter uses the same error code for
when profile banners POSTed to them can't be processed correctly, so it
seems like a borrow that is at least with precident.

The validation of the expected format only checks that we're getting a list of
things. I should validate that we have the expected entities
(user, area, x, y, z, button) in each element but that can wait.